refactor(test): dedupe retry server helper and simplify final send
