+++ src/executor.rs
@@ -978,61 +978,6 @@ mod tests {
     use super::*;
     use crate::discovery::{JsonSchema, JsonSchemaProperty, RestDescription, RestMethod};
     use serde_json::json;
-    use std::sync::{
-        atomic::{AtomicUsize, Ordering},
-        Arc,
-    };
-    use tokio::io::{AsyncReadExt, AsyncWriteExt};
-    use tokio::net::TcpListener;
-
-    fn reason_phrase(code: u16) -> &'static str {
-        match code {
-            200 => "OK",
-            429 => "Too Many Requests",
-            500 => "Internal Server Error",
-            _ => "Status",
-        }
-    }
-
-    pub(super) async fn spawn_response_server(
-        responses: Vec<(u16, Option<u64>, &'static str)>,
-    ) -> (
-        String,
-        Arc<AtomicUsize>,
-        tokio::task::JoinHandle<Result<(), std::io::Error>>,
-    ) {
-        let listener = TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr = listener.local_addr().unwrap();
-        let hits = Arc::new(AtomicUsize::new(0));
-        let hits_clone = Arc::clone(&hits);
-
-        let handle = tokio::spawn(async move {
-            for (status, retry_after, body) in responses {
-                let (mut socket, _) = listener.accept().await?;
-                let mut buf = [0u8; 2048];
-                let _ = socket.read(&mut buf).await?;
-                hits_clone.fetch_add(1, Ordering::SeqCst);
-
-                let mut extra_headers = String::new();
-                if let Some(v) = retry_after {
-                    extra_headers.push_str(&format!("Retry-After: {v}\r\n"));
-                }
-
-                let response = format!(
-                    "HTTP/1.1 {} {}\r\nContent-Type: application/json\r\nContent-Length: {}\r\nConnection: close\r\n{}\r\n{}",
-                    status,
-                    reason_phrase(status),
-                    body.len(),
-                    extra_headers,
-                    body
-                );
-                socket.write_all(response.as_bytes()).await?;
-            }
-            Ok(())
-        });
-
-        (format!("http://{addr}/"), hits, handle)
-    }
 
     #[test]
     fn test_pagination_config_default() {
@@ -1794,9 +1739,11 @@ async fn test_execute_method_missing_path_param() {
 
 #[tokio::test]
 async fn test_execute_method_retries_on_429_in_generic_path() {
-    let (base, hits, handle) =
-        tests::spawn_response_server(vec![(429, Some(0), "{}"), (200, None, "{\"ok\":true}")])
-            .await;
+    let (base, hits, handle) = crate::test_utils::spawn_response_server(vec![
+        crate::test_utils::mock_http_response(429, Some(0), "{}"),
+        crate::test_utils::mock_http_response(200, None, "{\"ok\":true}"),
+    ])
+    .await;
 
     let doc = RestDescription {
         base_url: Some(base),
@@ -1836,7 +1783,13 @@ async fn test_execute_method_retries_on_429_in_generic_path() {
 
 #[tokio::test]
 async fn test_execute_method_post_does_not_retry_on_429() {
-    let (base, hits, handle) = tests::spawn_response_server(vec![(429, Some(0), "{}")]).await;
+    let (base, hits, handle) =
+        crate::test_utils::spawn_response_server(vec![crate::test_utils::mock_http_response(
+            429,
+            Some(0),
+            "{}",
+        )])
+        .await;
 
     let doc = RestDescription {
         base_url: Some(base),
